requirements.txt
google-generativeai
pydub
moviepy==1.0.3
PyPDF2
python-dotenv
audioop-lts
jinja2
fpdf2